Create electron-forge configuration and update scripts, dependencies in package.json
This commit adds a new forge.config.js file to configure electron-forge with various plugins and makers. The start and test scripts in package.json are also updated to use electron-forge and two new scripts are added for package management. Also, additional devDependencies related to electron-forge and a new dependency 'electron-squirrel-startup' are added. It also will now ignore the 'out' directory as defined in the updated .gitignore, and includes updates to package-lock.json.